LlmExperiments/ImageGeneration/pyproject.toml
[project]
name = "image-generation"
version = "0.1.0"
description = "Generate images via Gemini and ChatGPT/OpenAI models from prompt files."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-genai",
  "openai",
  "pillow",
]

[tool.uv]
dev-dependencies = []

